- Apply complete codebase from MultiFusion branch
- Add AI MultiFusion features for advanced commit analysis
- Upgrade dashboard analytics with comprehensive new functionalities
- Improve sync manager with WebSocket support and real-time updates
- Add intelligent task assignment system
- Enhance frontend with modern React components and UI/UX improvements
- Integrate HAN model and MultiFusion model for AI-powered analysis
- Improve database schema with enhanced migrations and data models
- Add comprehensive testing, monitoring, and error handling
- Remove outdated components and consolidate codebase structure

+++ backend/db/models/pull_requests.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Table, Column, Integer, String, DateTime, func, ForeignKey
+from sqlalchemy import Table, Column, Integer, String, DateTime, func, ForeignKey, BigInteger
 from db.metadata import metadata
 
 pull_requests = Table(
     'pull_requests',
     metadata,
     Column('id', Integer, primary_key=True, autoincrement=True),
-    Column('github_id', Integer, nullable=True),
+    Column('github_id', BigInteger, nullable=True),  # Use BigInteger for GitHub IDs
     Column('title', String(255), nullable=False),
     Column('description', String(255), nullable=True),
     Column('state', String(50), nullable=True),
